Implements vote

+++ app/src/main/java/com/wizeline/brainstormingapp/App.kt
@@ -39,7 +39,7 @@ class App : Application() {
                 return Single.just(listOf(Message("mockId", "mockRoomId", "aRemote@email.com", "Mock text")))
             }
 
-            override fun vote(messageId: String, vote: Long): Single<Vote> {
+            override fun vote(messageId: String, vote: Int): Single<Vote> {
                 return Single.just(Vote("anId", "aMessageId", "aVoter@email.com", 1))
             }
         }
