+++ app/src/main/java/com/wizeline/brainstormingapp/repository/Repository.kt
@@ -20,6 +20,6 @@ interface Repository {
 
     fun getTopMessages(): Single<List<Message>>
 
-    fun vote(messageId: String, vote: Long): Single<Vote>
+    fun vote(messageId: String, vote: Int): Single<Vote>
 
 }
